requirements.txt
pandas
numpy
ta
textblob
matplotlib
gym
stable-baselines3
MetaTrader5
python-dotenv
optuna
flake8
pytest